# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "SanPham"
                ],
                "summary": "Lấy danh sách sản phẩm",
                "description": "Trả về danh sách tất cả các sản phẩm.",
                "operationId": "816847c99722b76c6ac4993a8f0c580c",
                "responses": {
                    "200": {
                        "description": "Dữ liệu được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu được lấy thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/products/sanPhamTheoDM/{MaDanhMuc}": {
            "get": {
                "tags": [
                    "SanPham"
                ],
                "summary": "Lấy danh sách sản phẩm theo mã danh mục",
                "description": "Trả về danh sách sản phẩm thuộc một danh mục cụ thể, chỉ khi danh mục có parent_id.",
                "operationId": "f18b87b1ea7fc5731eb889f8d56c8d44",
                "parameters": [
                    {
                        "name": "MaDanhMuc",
                        "in": "path",
                        "description": "Mã danh mục cần lấy sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu được lấy thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Danh mục không tồn tại hoặc không có parent_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Danh mục không tồn tại hoặc không có parent_id"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi máy chủ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi máy chủ"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/{MaSP}": {
            "get": {
                "tags": [
                    "SanPham"
                ],
                "summary": "Lấy thông tin chi tiết sản phẩm",
                "description": "Trả về thông tin chi tiết của sản phẩm dựa trên mã sản phẩm.",
                "operationId": "368797411e7b2674021f243264de6b9f",
                "parameters": [
                    {
                        "name": "MaSP",
                        "in": "path",
                        "description": "Mã sản phẩm cần lấy thông tin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy dữ liệu thành công"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi khi lấy dữ liệu"
                    },
                    "404": {
                        "description": "Sản phẩm không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "title": "Product Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Product Name"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a product description."
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}